+++ Tracker/CoreData/TrackerCategoryStore/TrackerCategoryStore.swift
@@ -1,29 +1,9 @@
 import UIKit
 import CoreData
 
-protocol TrackerCategoryStoreDelegate: AnyObject {
-    func store(_ store: TrackerCategoryStore, didUpdate update: TrackerCategoryStoreUpdate)
-}
-
-struct TrackerCategoryStoreUpdate {
-    struct Move: Hashable {
-        let oldIndex: Int
-        let newIndex: Int
-    }
-    let insertedIndexes: IndexSet
-    let deletedIndexes: IndexSet
-    let updatedIndexes: IndexSet
-    let movedIndexes: Set<Move>
-}
-
-enum TrackerCategoryStoreError: Error {
-    case decodingErrorInvalidTitle
-    case trackerCategoryNotFound
-    case categoryAlreadyExists
-}
-
 final class TrackerCategoryStore: NSObject {
     
+    // MARK: - Properties
     private let context: NSManagedObjectContext
     private var fetchedResultsController: NSFetchedResultsController<TrackerCategoryCoreData>?
     
@@ -33,37 +13,40 @@ final class TrackerCategoryStore: NSObject {
     private var updatedIndexes: IndexSet?
     private var movedIndexes: Set<TrackerCategoryStoreUpdate.Move>?
     
-    
+    // MARK: - Init
     convenience override init() {
-        let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
-        try! self.init(context: context)
+        guard let delegate = UIApplication.shared.delegate as? AppDelegate
+        else {
+            assertionFailure("❌ Unable to cast AppDelegate")
+            do {
+                try self.init(context: TrackerCategoryStore.dummyContext())
+            } catch {
+                fatalError("❌ Critical: Cannot initialize even with dummy context: \(error)")
+            }
+            return
+        }
+
+        let context = delegate.persistentContainer.viewContext
+
+        do {
+            try self.init(context: context)
+        } catch {
+            assertionFailure("❌ Failed to initialize TrackerCategoryStore: \(error)")
+            do {
+                try self.init(context: TrackerCategoryStore.dummyContext())
+            } catch {
+                fatalError("❌ Critical: Cannot initialize even with dummy context: \(error)")
+            }
+        }
     }
-    
+
     init(context: NSManagedObjectContext) throws {
         self.context = context
         super.init()
         try setupFetchedResultsController()
     }
     
-    private func setupFetchedResultsController() throws {
-        let fetchRequest = TrackerCategoryCoreData.fetchRequest()
-        
-        fetchRequest.sortDescriptors = [
-            NSSortDescriptor(key: "title", ascending: true)
-        ]
-        
-        let controller = NSFetchedResultsController(
-            fetchRequest: fetchRequest,
-            managedObjectContext: context,
-            sectionNameKeyPath: nil,
-            cacheName: nil
-        )
-        
-        controller.delegate = self
-        self.fetchedResultsController = controller
-        try controller.performFetch()
-    }
-    
+    // MARK: - Public Functions
     func addTrackerCategory(_ trackerCategory: TrackerCategory) throws {
         let trackerCategoryCoreData = TrackerCategoryCoreData(context: context)
         trackerCategoryCoreData.title = trackerCategory.title
@@ -137,6 +120,45 @@ final class TrackerCategoryStore: NSObject {
         return categories.first
     }
     
+    
+    
+    // MARK: - Private Functions
+    private func setupFetchedResultsController() throws {
+        let fetchRequest = TrackerCategoryCoreData.fetchRequest()
+        
+        fetchRequest.sortDescriptors = [
+            NSSortDescriptor(key: "title", ascending: true)
+        ]
+        
+        let controller = NSFetchedResultsController(
+            fetchRequest: fetchRequest,
+            managedObjectContext: context,
+            sectionNameKeyPath: nil,
+            cacheName: nil
+        )
+        
+        controller.delegate = self
+        self.fetchedResultsController = controller
+        try controller.performFetch()
+    }
+    
+    private static func dummyContext() -> NSManagedObjectContext {
+        let container = NSPersistentContainer(name: "Data")
+        
+        // Используем in-memory store (данные не сохраняются)
+        let description = NSPersistentStoreDescription()
+        description.type = NSInMemoryStoreType
+        container.persistentStoreDescriptions = [description]
+        
+        container.loadPersistentStores { _, error in
+            if let error = error {
+                print("⚠️ Dummy context failed to load: \(error)")
+            }
+        }
+        
+        return container.viewContext
+    }
+
     private func categoryExists(with title: String) throws -> Bool {
         let fetchRequest: NSFetchRequest<TrackerCategoryCoreData> = TrackerCategoryCoreData.fetchRequest()
         fetchRequest.predicate = NSPredicate(format: "title == %@", title)
@@ -147,6 +169,7 @@ final class TrackerCategoryStore: NSObject {
     }
 }
 
+// MARK: - Extension
 extension TrackerCategoryCoreData {
     func toTrackerCategory() -> TrackerCategory {
         let trackersArray = (trackers?.allObjects as? [TrackerCoreData])?.map { $0.toTracker() } ?? []
@@ -157,6 +180,7 @@ extension TrackerCategoryCoreData {
     }
 }
 
+// MARK: - Extension NSFetchedResultsControllerDelegate
 extension TrackerCategoryStore: NSFetchedResultsControllerDelegate {
     func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
         insertedIndexes = IndexSet()
