+++ Tracker/CoreData/TrackerCategoryStore/TrackerCategoryStoreDelegate.swift
@@ -1,0 +1,3 @@
+protocol TrackerCategoryStoreDelegate: AnyObject {
+    func store(_ store: TrackerCategoryStore, didUpdate update: TrackerCategoryStoreUpdate)
+}
