+++ Tracker/CoreData/TrackerCategoryStore/TrackerCategoryStoreError.swift
@@ -1,0 +1,5 @@
+enum TrackerCategoryStoreError: Error {
+    case decodingErrorInvalidTitle
+    case trackerCategoryNotFound
+    case categoryAlreadyExists
+}
