+++ Tracker/CoreData/TrackerRecordStore/TrackerRecordStoreDelegate.swift
@@ -1,0 +1,3 @@
+protocol TrackerRecordStoreDelegate: AnyObject {
+    func store(_ store: TrackerRecordStore, didUpdate update: TrackerRecordStoreUpdate)
+}
