+++ Tracker/CoreData/TrackerRecordStore/TrackerRecordStoreError.swift
@@ -1,0 +1,4 @@
+enum TrackerRecordStoreError: Error {
+    case decodingErrorInvalidId
+    case decodingErrorInvalidDate
+}
