+++ Tracker/CoreData/TrackerRecordStore/TrackerRecordStoreUpdate.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+struct TrackerRecordStoreUpdate {
+    struct Move: Hashable {
+        let oldIndex: Int
+        let newIndex: Int
+    }
+    let insertedIndexes: IndexSet
+    let deletedIndexes: IndexSet
+    let updatedIndexes: IndexSet
+    let movedIndexes: Set<Move>
+}
