+++ Tracker/CoreData/TrackerStore/TrackerStoreDelegate.swift
@@ -1,0 +1,3 @@
+protocol TrackerStoreDelegate: AnyObject {
+    func store(_ store: TrackerStore, didUpdate update: TrackerStoreUpdate)
+}
