+++ Tracker/CoreData/TrackerStore/TrackerStoreError.swift
@@ -1,0 +1,8 @@
+enum TrackerStoreError: Error {
+    case decodingErrorInvalidColor
+    case decodingErrorInvalidId
+    case decodingErrorInvalidEmoji
+    case decodingErrorInvalidName
+    case decodingErrorInvalidSchedule
+    case trackerNotFound
+}
