+++ Tracker/CoreData/TrackerStore/TrackerStoreUpdate.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+struct TrackerStoreUpdate {
+    struct Move: Hashable {
+        let oldIndex: Int
+        let newIndex: Int
+    }
+    let insertedIndexes: IndexSet
+    let deletedIndexes: IndexSet
+    let updatedIndexes: IndexSet
+    let movedIndexes: Set<Move>
+}
